+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `init` convenience command will now populate the working directory with files from the starter project
 
+### Changed
+
+- `secrets set` no longer prompts twice for confirmation if the secret set does not exist
+
 ## [0.0.9] - 2025-02-27
 
 ### Added